html/semantics/embedded-content-0/the-audio-element/audio_audioTracks_return_type.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Audio Test: audioTracks property of audio element</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://www.w3.org/TR/html5/embedded-content-0.html#htmlmediaelement">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<audio id="audio1">
  Your browser does not support HTML5 audio.
</audio>
<script>

test(function() {
  var a = document.getElementById('audio1');
  assert_equals(Object.prototype.toString.call(a.audioTracks), "[object AudioTrackList]", "The return type of audioTracks attribute");
}, "Check if the audioTracks attribute of a audio element must return AudioTrackList object");

</script>
